update: isolate made and wode from core jobs
