+++ src/py/mat3ra/notebooks_utils/core/entity/job/api.py
@@ -2,8 +2,6 @@ import urllib.request
 from typing import List, Optional, Union
 
 from mat3ra.api_client import APIClient, JobEndpoints
-from mat3ra.made.material import Material
-from mat3ra.wode import Workflow
 
 
 def save_files(job_id: str, job_endpoint: JobEndpoints, filename_on_cloud: str, filename_on_disk: str) -> None:
@@ -41,67 +39,46 @@ def get_jobs_statuses_by_ids(endpoint: JobEndpoints, job_ids: List[str]) -> List
 
 def create_job(
     api_client: APIClient,
-    materials: List[Union[dict, Material]],
-    workflow: Union[dict, Workflow],
+    material_dicts: List[dict],
+    job_workflow_dict: dict,
     project_id: str,
     owner_id: str,
     prefix: str,
     compute: Optional[dict] = None,
 ) -> Union[dict, List[dict]]:
     """
-    Creates jobs for each material using an embedded workflow with any context already applied.
+    Creates jobs using pre-serialised material and workflow dicts.
 
     Args:
         api_client (APIClient): API client instance carrying the authorization context.
-        materials (list): List of material dicts or mat3ra-made Material objects.
-        workflow: Workflow dict or Workflow object with important settings already applied.
+        material_dicts (list[dict]): Serialised material dicts.
+        job_workflow_dict (dict): Serialised workflow dict.
         project_id (str): Project ID.
         owner_id (str): Account ID.
         prefix (str): Job name prefix.
         compute (dict, optional): Compute configuration dict.
 
     Returns:
-        list[dict]: List of created job dicts.
+        dict | list[dict]: Created job(s).
     """
-    material_dicts = [m.to_dict() if isinstance(m, Material) else m for m in materials]
-    job_workflow_dict = workflow.to_dict() if isinstance(workflow, Workflow) else workflow
     job_workflow_dict.pop("_id", None)
     is_multimaterial = job_workflow_dict.get("isMultiMaterial", False)
 
-    config = {
+    config: dict = {
         "_project": {"_id": project_id},
         "workflow": job_workflow_dict,
         "owner": {"_id": owner_id},
         "name": prefix,
+        "_material": {"_id": material_dicts[0]["_id"]},
     }
 
     if is_multimaterial:
-        config["_material"] = {"_id": material_dicts[0]["_id"]}
         config["_materials"] = [{"_id": m["_id"]} for m in material_dicts]
-    else:
-        config["_material"] = {"_id": material_dicts[0]["_id"]}
 
     if compute:
         config["compute"] = compute
+
     return api_client.jobs.create(config)
-
-
-def get_convergence_series(client: APIClient, job_id: str, subworkflow_index: int = 0) -> List[dict]:
-    """
-    Returns the convergence series from a finished convergence job.
-
-    Args:
-        client: API client instance.
-        job_id: ID of the finished convergence job.
-        subworkflow_index: Index of the convergence subworkflow (default 0).
-
-    Returns:
-        List of dicts with keys "x", "parameter", "y".
-    """
-    finished_job = client.jobs.get(job_id)
-    job_workflow = Workflow.create(finished_job["workflow"])
-    subworkflow = job_workflow.subworkflows[subworkflow_index]
-    return subworkflow.convergence_series(finished_job.get("scopeTrack"))
 
 
 def submit_jobs(endpoint: JobEndpoints, job_ids: List[str]) -> None:
